+++ src/main/java/org/truffleruby/core/module/ModuleNodes.java
@@ -2286,6 +2286,48 @@ public abstract class ModuleNodes {
 
     }
 
+    @CoreMethod(names = "set_temporary_name", required = 1)
+    public abstract static class SetTemporaryNameNode extends CoreMethodArrayArgumentsNode {
+
+        @TruffleBoundary
+        @Specialization
+        RubyModule setTemporaryName(RubyModule self, Object name,
+                @Cached ToJavaStringNode toJavaStringNode) {
+            if (name == nil) {
+                self.fields.setTemporaryName(null);
+                return self;
+            }
+
+            final String string = toJavaStringNode.execute(this, name);
+            validateName(string, self);
+            self.fields.setTemporaryName(string);
+
+            return self;
+        }
+
+        private void validateName(String name, RubyModule self) {
+            if (name.isEmpty()) {
+                throw new RaiseException(
+                        getContext(this),
+                        coreExceptions(this).argumentError("empty class/module name", this));
+            }
+
+            if (self.fields.hasFullName()) {
+                throw new RaiseException(
+                        getContext(this),
+                        coreExceptions(this).runtimeError("can't change permanent name", this));
+            }
+
+            if (Identifiers.isValidConstantPath(name)) {
+                throw new RaiseException(
+                        getContext(this),
+                        coreExceptions(this).argumentError(
+                                "the temporary name must not be a constant path to avoid confusion", this));
+            }
+        }
+
+    }
+
     @GenerateUncached
     @CoreMethod(names = "using", required = 1, visibility = Visibility.PRIVATE, alwaysInlined = true)
     public abstract static class ModuleUsingNode extends UsingNode {